pyproject.toml
[tool.poetry]
name = "mosparser"
version = "0.1.0"
description = ""
authors = ["Tvister7 <kirzap123@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
psycopg2-binary = "^2.9.3"
python-dotenv = "^0.19.2"
beautifulsoup4 = "^4.10.0"
requests = "^2.27.1"
sqlmodel = "^0.0.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
